backend/requirements.txt
flask
Flask-SQLAlchemy
pymysql
scikit-learn
pandas
numpy
Flask-Cors
marshmallow
joblib